extract audio section widget
